fastapi/requirements.txt
fastapi==0.68.0
torchvision==0.13.0
torch==1.12.0
python-multipart==0.0.5
aiofiles==0.4.0
joblib==1.2.0
matplotlib==3.6.3
natsort==8.2.0
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
requests==2.27.1
scikit_learn==1.2.0
six==1.16.0
tensorflow==2.11.0

